requirements.txt
# LeadTool Requirements
# Unified Lead Generation and Management System

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter
# sqlite3 is built-in with Python, no need to install

# Web Scraping
scrapy>=2.11.0
scrapy-playwright>=0.0.33
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Dashboard
streamlit==1.28.1
plotly==5.17.0
pandas>=2.2.0
numpy>=1.26.0

# Scheduler
apscheduler==3.10.4
croniter==1.4.1

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.9
python-multipart==0.0.6

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Logging
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
supervisor==4.2.5

# Optional: For advanced features
celery==5.3.4  # For distributed task queue
redis==5.0.1   # For Celery broker
elasticsearch==8.11.0  # For advanced search
